requirements.txt
torch>=2.1
torchvision>=0.16
ultralytics>=8.2
opencv-python>=4.8
numpy>=1.24
tqdm>=4.66
pyyaml>=6.0
scipy>=1.11
